pyproject.toml
[tool.poetry]
name = "mcc-import"
version = "0.1.0"
description = "Import tool for \"My city's climate\""
authors = ["Ivan Kavaliou <info@ikovalev.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
pymongo = {extras = ["srv"], version = "^4.0.1"}
click = "^8.0.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
import = "import"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
